+++ src/main/java/com/example/ged/src/user/models/UserInfo.java
@@ -1,11 +1,8 @@
 package com.example.ged.src.user.models;
 
 import com.example.ged.config.BaseEntity;
-import com.example.ged.src.user.models.LoginMethod;
-import lombok.AccessLevel;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 
@@ -18,7 +15,7 @@ public class UserInfo extends BaseEntity {
     @Id
     @Column(name="userIdx",nullable = false,updatable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long userIdx;
+    private Integer userIdx;
 
     @Column(name="userName",nullable = false, length = 10)
     private String userName;
@@ -57,44 +54,5 @@ public class UserInfo extends BaseEntity {
         this.backgroundImageUrl = backgroundImageUrl;
         this.socialId = socialId;
     }
-//    @Id
-//    @Column(name="userIdx",nullable = false,updatable = false)
-//    @GeneratedValue(strategy = GenerationType.IDENTITY)
-//    private Integer userIdx;
-//
-//    @Column(name="socialId",nullable = false)
-//    private String socialId;
-//
-//    @Column(name="name",nullable = false)
-//    private String name;
-//
-//    @Column(name="email",nullable = false)
-//    private String email;
-//
-//    @Column(name="profileImageUrl",columnDefinition = "TEXT")
-//    private String profileImageUrl;
-//
-//    @Column(name="introduce")
-//    private String introduce;
-//
-//    @Column(name="snsUrl",columnDefinition = "TEXT")
-//    private String snsUrl;
-//
-//    @Column(name="loginMethod",nullable = false)
-//    @Enumerated(EnumType.STRING)
-//    private LoginMethod loginMethod;
-//
-//    @Column(name="status",nullable = false)
-//    private String status = "ACTIVE";
-//
-//    private UserInfo(Integer userIdx,String socialId,String name,String email,String profileImageUrl,String introduce,String snsUrl,LoginMethod loginMethod){
-//        this.userIdx = userIdx;
-//        this.socialId = socialId;
-//        this.name = name;
-//        this.email = email;
-//        this.profileImageUrl = profileImageUrl;
-//        this.introduce = introduce;
-//        this.snsUrl = snsUrl;
-//        this.loginMethod = loginMethod;
-//    }
+
 }
